requirements.txt
opencv-python
lxml
tqdm
tensorflow
absl-py
easydict
matplotlib
pillow
numpy